made hosted_upload into hosted util and hosted_upload main, standardized logging with logging util.
